pyproject.toml
[tool.poetry]
name = "alpha-trader"
version = "0.1.0"
description = ""
authors = ["hinata356 <piano.natsu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
alpha-vantage = "^2.3.1"
pandas = "^2.2.2"
snowflake-connector-python = "^3.11.0"
snowflake-id = "^1.0.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
